# api/openapi.yaml
openapi: 3.1.0
info:
  title: CloudBank API
  version: 0.1.0
  description: >
    HTTP API for CloudBank, a self-hosted web port of HomeBank. This document is
    the single source of truth for the contract: the Go server interfaces and
    the TypeScript client types are both generated from it. Resources are added
    milestone by milestone; this initial version covers the health and ping
    endpoints established by the server skeleton.
  license:
    name: AGPL-3.0-or-later
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: /
paths:
...
  /api/v1/admin/backup:
    get:
      operationId: hotBackup
      summary: Download a consistent copy of the whole database (admin only)
      responses:
        "200": { description: A SQLite database file., content: { application/octet-stream: { schema: { type: string, format: binary } } } }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/v1/backup/restore:
    post:
      operationId: restoreBackup
      summary: Restore a wallet backup JSON into a brand-new wallet
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: true }
      responses:
        "201": { description: The wallet was restored., content: { application/json: { schema: { type: object, properties: { walletId: { type: integer, format: int64 } }, required: [walletId] } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
...
  /api/v1/catalog/currencies:
    get:
      operationId: getCurrencyCatalog
      summary: The ISO 4217 currency catalog
      responses:
        "200":
          description: Catalog entries.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CatalogEntry"
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /api/v1/import/xhb:
    post:
      operationId: importXHB
      summary: Import a HomeBank .xhb file into a new wallet owned by the current user
      description: >
        The raw .xhb XML is sent as the request body. The import is atomic: a fresh
        wallet is created and populated with the file's currencies, accounts, payees,
        categories, tags, transactions (with splits and re-paired transfers),
        budgets, assignments, templates and schedules.
      requestBody:
        required: true
        content:
          application/xml:
            schema: { type: string, format: binary }
      responses:
        "201": { description: The file was imported., content: { application/json: { schema: { $ref: "#/components/schemas/ImportResult" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
...
components:
  responses:
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Authentication required.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Administrator access required.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
          required: [code, message]
      required: [error]
...
    CatalogEntry:
      type: object
      properties:
        code: { type: string }
        name: { type: string }
        symbol: { type: string }
        fracDigits: { type: integer }
        symbolPrefix: { type: boolean }
      required: [code, name, symbol, fracDigits, symbolPrefix]
...
    ImportResult:
      type: object
      properties:
        walletId: { type: integer, format: int64, description: "the newly created wallet" }
        counts: { type: object, additionalProperties: { type: integer }, description: "imported entity counts keyed by type (accounts, transactions, ...)" }
        warnings: { type: array, items: { type: string } }
      required: [walletId, counts, warnings]
...